Major Rewrite and Tone Test for NFAs (2nd Attempt)
(2nd attempt with minor markdown issues sorted)

We're re-toning the NFA structure to be more intent driven.  I've completed one example of this - Funding.  That's now open for review.

It's likely this format will go through a few iterations, so it's not yet canonical.

NB, I haven't changed any other NFAs, but they are no long valid as they'll need to be rewritten in the style we settle on.