Removed stagegates, QA authority and dependencies
On reflection this is not the right point in the process to be getting into stagegates, QA and dependencies.  Here's my thinking:

Stagegates needs to be worked through with the guy responding with an Idea.  And why does it matter at the Request stage anyway?  It should be a pure request, but this is stepping into the Project stage.

QA Authority should be assigned at the Request:Idea lock Contract.  Otherwise as an Idea submitter why should I just accept your QA.  Maybe I've got a better one to propose.

Dependencies - again, stepping into solutioneering.  Needs to be pure problem.