Moving this out of drafts as with Idea Template
This is probably not the end destination, but it's definitely not nfa/drafts and need to refer to it from [Pool Process.md](https://github.com/dreamcatcher-tech/dreamcatcher-tech.github.io/new/master/website/pool) so moving it here.  

Reasoning is that this is actually an Output, but as we're bootstrapping it already exists, and is needed in the process definition of the thing that will define it's location.

Happy to move it elsewhere, but bootstrapping, and this is closer.